+++ DKPlugins/DK/DKWindows.h
@@ -72,10 +72,10 @@ public:
 	static bool MiddleRelease();
 	static bool PhysicalMemory(unsigned long long& physicalMemory);
 	static bool PhysicalMemoryUsed(unsigned long long& physicalMemory);
-	static bool PhysicalMemoryUsedByApp(unsigned int& physicalMemory);
-	static bool PressKey(int key);
+	static bool PhysicalMemoryUsedByApp(size_t& physicalMemory);
+	static bool PressKey(WORD key);
 	static bool RefreshWindowsEnvironment();
-	static bool ReleaseKey(int key);
+	static bool ReleaseKey(WORD key);
 	static bool RightClick();
 	static bool RightPress();
 	static bool RightRelease();
@@ -88,12 +88,12 @@ public:
 	static bool SetTitle(const DKString& title);
 	static bool SetVolume(int& percent);
 	static bool Sleep(int milliseconds);
-	static bool StrokeKey(const int& key);
+	static bool StrokeKey(const WORD& key);
 	static bool TurnOffMonitor();
 	static bool TurnOnMonitor();
 	static bool VirtualMemory(unsigned long long& virtualMemory);
 	static bool VirtualMemoryUsed(unsigned long long& virtualMemory);
-	static bool VirtualMemoryUsedByApp(unsigned int& virtualMemory);
+	static bool VirtualMemoryUsedByApp(size_t& virtualMemory);
 	static bool WaitForImage(const DKString& file, int timeout);
 	static bool WheelDown();
 	static bool WheelUp();
